refactor: move RepositoryService into lib/registry/
