+++ src/components/dialogs/PackageDetailModal.tsx
