+++ src/components/dialogs/VulnerabilityDetailModal.tsx
