+++ src/components/dialogs/bulk-scan-modal.tsx
