+++ src/components/dialogs/delete-image-dialog.tsx
