+++ src/components/dialogs/index.ts
@@ -1,0 +1,10 @@
+export { AddRepositoryDialog } from './add-repository-dialog';
+export { BulkScanModal } from './bulk-scan-modal';
+export { CveClassificationDialog } from './cve-classification-dialog';
+export { DeleteImageDialog } from './delete-image-dialog';
+export { ExportImageDialogEnhanced } from './export-image-dialog-enhanced';
+export { NewScanModal } from './new-scan-modal';
+export { VulnerabilityDetailsModal } from './vulnerability-details-modal';
+export { VulnerabilitySelectionModal } from './vulnerability-selection-modal';
+export { PackageDetailModal } from './PackageDetailModal';
+export { VulnerabilityDetailModal } from './VulnerabilityDetailModal';
