+++ src/components/dialogs/new-scan-modal.tsx
