+++ src/components/dialogs/vulnerability-details-modal.tsx
