+++ src/lib/registry/RepositoryService.ts
