+++ src/lib/scanner/DatabaseAdapter.ts
@@ -2,7 +2,7 @@ import { prisma } from '@/lib/prisma';
 import { inspectDockerImage } from '@/lib/docker';
 import { IDatabaseAdapter, ScanReports } from './types';
 import type { ScanRequest } from '@/types';
-import { RepositoryService } from '@/services/RepositoryService';
+import { RepositoryService } from '@/lib/registry/RepositoryService';
 import { RegistryProviderFactory } from '@/lib/registry/providers/RegistryProviderFactory';
 import type { Repository } from '@/generated/prisma';
 import { createOrUpdateScanMetadata, saveScannerResultTables } from './result-savers';
